Added combat scripts, updated controller scripts, added the ability to pause
Added the weapon scripts for the combat.

The player now uses the new input system by having the script PlayerInputs use the singleton pattern and letting properties point to actions.

Added the ability to pause by making a singleton pause handler, also added a script mytime to keep track of the time when the game has not been paused